Updated the plugin description.
After the file generation open the generated file.
If the class is already the one we have generated before do not generate an additional class.

Implemented so that nature for the project setting is only visible when clicked on the project when it is not set, and then nowhere else after it is set.